requirements-phase5.txt
# Phase 5: Advanced Production Features Dependencies

# Core dependencies
asyncio>=3.4.3
typing-extensions>=4.0.0
dataclasses-json>=0.5.0
pydantic>=2.0.0

# Analytics and Business Intelligence
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Database and Storage
psycopg2-binary>=2.9.0
redis>=4.5.0
sqlalchemy>=2.0.0

# Vector and Graph Databases
qdrant-client>=1.6.0
neo4j>=5.0.0
sentence-transformers>=2.2.0

# Machine Learning and NLP
torch>=2.0.0
transformers>=4.30.0
nltk>=3.8.0

# Enterprise Authentication
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-ldap>=3.4.0
pyotp>=2.8.0
qrcode>=7.4.0

# Security and Encryption
cryptography>=41.0.0
bcrypt>=4.0.0
secrets>=3.3.0

# System Monitoring and Metrics
psutil>=5.9.0
prometheus-client>=0.17.0

# Cloud and Backup
boto3>=1.28.0
google-cloud-storage>=2.10.0
azure-storage-blob>=12.17.0

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
locust>=2.15.0
requests>=2.31.0

# Docker and Containerization
docker>=6.1.0
kubernetes>=27.2.0

# Chaos Engineering
chaos-toolkit>=1.14.0
litmus-python>=1.0.0

# High Availability
haproxy-stats>=2.0.0

# Configuration and Utilities
pyyaml>=6.0.0
jinja2>=3.1.0
click>=8.1.0
rich>=13.0.0

# Development and Debugging
ipython>=8.14.0
jupyter>=1.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0